mt(1): Change c_code to u_long because it has to hold an ioctl command.
Forgotten porting to x86_64.

+++ usr.bin/mt/mt.c
@@ -66,7 +66,7 @@
 
 static struct commands {
 	char *c_name;
-	int c_code;
+	u_long c_code;
 	int c_ronly;
 #if defined(__DragonFly__)
 	int c_flags;
